requirements.txt
requests==2.32.3
aiohttp==3.9.5
python-dotenv==1.0.1
pandas==2.2.2
tqdm==4.66.4